+++ tls/s2n_tls.h
@@ -40,8 +40,6 @@ extern int s2n_server_cert_send(struct s2n_connection *conn);
 extern int s2n_server_cert_recv(struct s2n_connection *conn);
 extern int s2n_server_status_send(struct s2n_connection *conn);
 extern int s2n_server_status_recv(struct s2n_connection *conn);
-extern int s2n_server_cert_verify_send(struct s2n_connection *conn);
-extern int s2n_server_cert_verify_recv(struct s2n_connection *conn);
 extern int s2n_server_key_send(struct s2n_connection *conn);
 extern int s2n_server_key_recv(struct s2n_connection *conn);
 extern int s2n_cert_req_recv(struct s2n_connection *conn);
@@ -56,6 +54,8 @@ extern int s2n_client_key_send(struct s2n_connection *conn);
 extern int s2n_client_key_recv(struct s2n_connection *conn);
 extern int s2n_client_cert_verify_recv(struct s2n_connection *conn);
 extern int s2n_client_cert_verify_send(struct s2n_connection *conn);
+extern int s2n_tls13_cert_verify_recv(struct s2n_connection *conn);
+extern int s2n_tls13_cert_verify_send(struct s2n_connection *conn);
 extern int s2n_server_nst_send(struct s2n_connection *conn);
 extern int s2n_server_nst_recv(struct s2n_connection *conn);
 extern int s2n_ccs_send(struct s2n_connection *conn);
